Removed Model.ts and renamed GroqModel.ts to Model.ts
